+++ wavmio/higher.go
@@ -70,7 +70,7 @@ func AdvanceInboxMessage() {
 func ResolveTypedPreimage(ty arbutil.PreimageType, hash common.Hash) ([]byte, error) {
 	return readBuffer(func(offset uint32, buf unsafe.Pointer) uint32 {
 		hashUnsafe := unsafe.Pointer(&hash[0])
-		return resolveTypedPreimage(uint8(ty), hashUnsafe, offset, buf)
+		return resolveTypedPreimage(uint32(ty), hashUnsafe, offset, buf)
 	}), nil
 }
